backend/requirements.txt
fastapi
uvicorn
openai-whisper
torch
transformers
librosa
gTTS
deep-translator
eng-to-ipa
